Rename `Error.ts` -^> `ErrorEvent.ts`
