+++ framework/lib/Events/Modules/ErrorEvent.ts
