+++ framework/lib/Events/Modules/index.ts
@@ -1,3 +1,3 @@
-export * from "./Error";
+export * from "./ErrorEvent";
 export * from "./InteractionCreateEvent";
 export * from "./ReadyEvent";
